+++ survey_percent_question/models/__init__.py
@@ -21,3 +21,5 @@
 ##############################################################################
 
 from . import survey
+from . import survey_input
+from . import survey_question
