+++ survey_percent_question/models/survey.py
@@ -21,101 +21,9 @@
 ##############################################################################
 
 from collections import defaultdict
-from decimal import Decimal, InvalidOperation
 
-from openerp import api, fields, models
+from openerp import api, models
 from openerp.tools.translate import _
-
-from openerp.addons.survey.survey import (
-    dict_keys_startswith,
-)
-
-
-def to_decimal(value):
-    if value:
-        value = value.strip().replace(",", ".")
-    else:
-        value = '0'
-
-    return Decimal(value)
-
-
-class SurveyInputLine(models.Model):
-    _inherit = 'survey.user_input_line'
-
-    @api.model
-    def get_old_lines(self, user_input_id, question):
-        """ Returns the old input lines """
-        return self.search([('user_input_id', '=', user_input_id),
-                            ('survey_id', '=', question.survey_id.id),
-                            ('question_id', '=', question.id)])
-
-    @api.model
-    def save_line_percent_split(self, user_input_id, question, post,
-                                answer_tag):
-        defaults = {
-            "answer_type": "number",
-            "user_input_id": user_input_id,
-            "question_id": question.id,
-            "page_id": question.page_id.id,
-            "survey_id": question.survey_id.id,
-            "skipped": False
-        }
-
-        old_lines = self.get_old_lines(user_input_id, question)
-        if old_lines:
-            old_lines.unlink()
-
-        answers = dict_keys_startswith(post, answer_tag)
-        comment_label = answers.pop(answer_tag + '_comment_label', None)
-        comment_value = answers.pop(answer_tag + '_comment_value', None)
-        if comment_label:
-            comment_value = to_decimal(comment_value)
-            self.create(dict(defaults, **{
-                "value_number": comment_value,
-                "value_text": comment_label,
-            }))
-
-        for row in question.labels_ids:
-            a_tag = "{0}_{1}".format(answer_tag, row.id)
-            self.create(dict(defaults, **{
-                "value_number": float(to_decimal(answers.get(a_tag))),
-                "value_suggested_row": row.id,
-            }))
-
-        return True
-
-
-class SurveyQuestion(models.Model):
-    _inherit = 'survey.question'
-
-    type = fields.Selection(
-        selection_add=[('percent_split', 'Percentage')],
-    )
-
-    @api.model
-    def validate_percent_split(self, question, post, answer_tag):
-        answers = dict_keys_startswith(post, answer_tag)
-        dec_values = []
-        comment_label = answers.pop(answer_tag + '_comment_label', None)
-        comment_value = answers.get(answer_tag + '_comment_value', None)
-        if comment_value and not comment_label:
-            return {answer_tag: _(
-                "You must provide a label for your extra value"
-            )}
-
-        for v in answers.values():
-            try:
-                dec_values.append(to_decimal(v))
-            except InvalidOperation:
-                return {answer_tag: _("{0} is not a number").format(v)}
-
-        if not sum(dec_values) == 100:
-            return {answer_tag: _(
-                'The sum of all values must be 100%, not {0}%'
-            ).format(sum(dec_values))}
-
-        return {}
 
 
 class Survey(models.Model):
